app: project initialization via create-react-app and adaptation of a force graph example
The example I adapted can be found at [0]. I wanted to have an working
example to implement a PoC of the graph layout switching idea I had for
VisHub's author network viewer.

Unfortunately, the example was not runnable out-of-the-box, which I
think is due to changes in different library versions but also in
Typescript 2.0.
Addtionally, I do not think that some things were implemented in a clear
way in this example.
Which is why I took the time to fix every build error and change things
until they made sense to me or somebody else with a preference for
'clearer' code.

First of all, the refs declared in {links, labels, nodes}.tsx were
broken and needed a more specific type to be correct, i.e.
SVGLineElement becomes React.RefObject<SVGLineElement>.
Then, I did not like setting their value in the render method, which is
why I refactored the initialization of the reference to be part of the
object initialization, i.e. ref: React.RefObject<SVGLineElement> =
React.createRef();

Then, the defined types d3{Node, Link, Graph} were slightly broken and
could not work in Typescript as written by the author in [0].
He implemented the interfaces needed for d3's simulation
(SimulationLinkDatum, SimulationNodeDatum) but added extra properties
(id/group for d3Node and value for d3Link) upon which he based his code.
The links were given an id which have been used in d3.forceLink().id()
which expects a class that implements SimulationNodeDatum.
Because SimulationNodeDatum has no id property, the types were not
equal.

The last, and biggest problem, was a type inference error:
In nodes.tsx:37, the types <SVGCircleElement, d3Types.d3Node> have to be
added manually to the select() and drag() calls such that Typescript is
able to correctly infer the types.

[0] https://spin.atomicobject.com/2017/07/20/d3-react-typescript/