components: rename Links/Nodes into Edges/Vertices
Links and Nodes were too overloaded for me in the context of web
development, i.e. URLs/URIs/<a>-tags and other stuff for Links and DOM
nodes for example for Nodes.
Using the terms Edges/Vertices, which atleast for me are nearer to the
mathematical and computer graphics definition, helped me.