components: _really_ implement unique keys this game
Did a mistake before to forget that I reuse one of the collections for
the labels, which means that 2 elements will always have the same key,
i.e. a collection occurs.
Also, my mistake was thinking that collection.name inside the function
would return the original name and not collection, meaning that all
caller would get the same key.

Now, a string is passed to the callee as a prefix which ensures the
uniqueness, instead of trying some weird reflection thing and failing at
it because there is no real first-class reflection support.